requirements.txt
langchain==0.3.25
langchain-google-genai==2.0.10
python-dotenv==1.0.0
pydantic==2.10.6
pandas>=2.2.0
pillow>=10.0.0